WIP: Launching correct container
For diagnostic purposes, the number of containers is reduced to four.
In the scope, the number of the container which will be used for each is
included in the title.

The database isn't always updated as much as it should be.  When a
container passes control off to another container, the second doesn't
get focus.